moved endpoint calls to super class
